requirements.txt
Pillow>=9.0.0
numpy>=1.22.0
colorama>=0.4.4